requirements.txt
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
numpy==1.21.0
pandas==1.3.0
matplotlib==3.4.2
scikit-learn==0.24.2
gunicorn==20.1.0